Add a reusable user avatar widget
